move all methods to `AccountIdRef`
