+++ src/serde.rs
@@ -19,7 +19,7 @@ impl<'de> de::Deserialize<'de> for AccountId {
         D: de::Deserializer<'de>,
     {
         let account_id = Box::<str>::deserialize(deserializer)?;
-        AccountId::validate(&account_id).map_err(|err| {
+        crate::validation::validate(&account_id).map_err(|err| {
             de::Error::custom(format!("invalid value: \"{}\", {}", account_id, err))
         })?;
         Ok(AccountId(account_id))
